\name{dist.HalfCauchy}
\alias{dhalfcauchy}
\title{Half-Cauchy Distribution}
\description{
  This is a C++ version of the \code{dhalfcauchy} function in the
  \pkg{LaplacesDemon} package.
}
\usage{
dhalfcauchy(x, scale=25, log=FALSE)
}
\arguments{
  \item{x}{See \code{dhalfcauchy}.}
  \item{scale}{See \code{dhalfcauchy}.}
  \item{log}{See \code{dhalfcauchy}.}
}
\details{
  See \code{dhalfcauchy}.
}
\value{
  See \code{dhalfcauchy}.
}
\seealso{
  \url{www.bayesian-inference.com/softwaredoc/dist.HalfCauchy}
}
\keyword{Distribution}
