\name{dist.Multivariate.Normal.Cholesky}
\alias{dmvnc}
\title{Multivariate Normal Distribution: Cholesky Parameterization}
\description{
  This is a C++ version of the \code{dmvnc} function in the
  \pkg{LaplacesDemon} package.
}
\usage{
dmvnc(x, mu, U, log=FALSE) 
}
\arguments{
  \item{x}{See \code{dmvnc}.}
  \item{mu}{See \code{dmvnc}.}
  \item{U}{See \code{dmvnc}.}
  \item{log}{See \code{dmvnc}.}
}
\details{
  See \code{dmvnc}.
}
\value{
  See \code{dmvnc}.
}
\author{Statisticat, LLC. \email{software@bayesian-inference.com}}
\seealso{
  \url{www.bayesian-inference.com/softwaredoc/dist.Multivariate.Normal.Cholesky}
}
\keyword{Distribution}