\name{dist.Multivariate.Normal.Precision.Cholesky}
\alias{rmvnpc}
\title{Multivariate Normal Distribution: Precision-Cholesky Parameterization}
\description{
  This is the C++ version of the \code{rmvnpc} function in the
  \pkg{LaplacesDemon} package.
}
\usage{
rmvnpc(n=1, mu=rep(0,ncol(U)), U)
}
\arguments{
  \item{n}{See \code{rmvnpc}.}
  \item{mu}{See \code{rmvnpc}.}
  \item{U}{See \code{rmvnpc}.}
}
\details{
  See \code{rmvnpc}.
}
\value{
  See \code{rmvnpc}.
}
\author{Statisticat, LLC. \email{software@bayesian-inference.com}}
\seealso{
  \url{www.bayesian-inference.com/softwaredoc/dist.Multivariate.Normal.Precision.Cholesky}
}
\keyword{Distribution}