\name{dist.Multivariate.Normal.Precision}
\alias{rmvnp}
\title{Multivariate Normal Distribution: Precision Parameterization}
\description{
  This is the C++ version of the \code{dmvnp} function in the
  \pkg{LaplacesDemon} package.
}
\usage{
rmvnp(n=1, mu=rep(0,ncol(Omega)), Omega)
}
\arguments{
  \item{n}{See \code{rmvnp}.}
  \item{mu}{See \code{rmvnp}.}
  \item{Omega}{See \code{rmvnp}.}
}
\details{
  See \code{rmvnp}.
}
\value{
  See \code{rmvnp}.
}
\author{Statisticat, LLC. \email{software@bayesian-inference.com}}
\seealso{
  \url{www.bayesian-inference.com/softwaredoc/dist.Multivariate.Normal.Precision}
}
\keyword{Distribution}