\name{dist.Multivariate.Normal}
\alias{dmvn}
\title{Multivariate Normal Distribution}
\description{
  This is a C++ version of the \code{dmvn} function in the
  \pkg{LaplacesDemon} package.
}
\usage{
dmvn(x, mu, Sigma, log=FALSE) 
}
\arguments{
  \item{x}{See \code{dmvn}.}
  \item{mu}{See \code{dmvn}.}
  \item{Sigma}{See \code{dmvn}.}
  \item{log}{See \code{dmvn}.}
}
\details{
  See \code{dmvn}.
}
\value{
  See \code{dmvn}.
}
\author{Statisticat, LLC. \email{software@bayesian-inference.com}}
\seealso{
  \url{www.bayesian-inference.com/softwaredoc/dist.Multivariate.Normal}
}
\keyword{Distribution}