\name{dist.Wishart.Cholesky}
\alias{dwishartc}
\alias{rwishartc}
\title{Wishart Distribution: Cholesky Parameterization}
\description{
  This is a C++ version of the \code{dwishart} and \code{rwishart}
  functions in the \pkg{LaplacesDemon} package.
}
\usage{
   dwishartc(U, nu, S, log=FALSE)
   rwishartc(nu, S)
}
\arguments{
  \item{U}{See \code{dwishartc}.}
  \item{nu}{See \code{dwishartc}.}
  \item{S}{See \code{dwishartc}.}
  \item{log}{See \code{dwishartc}.}
}
\details{
  See \code{dwishartc}.
}
\value{
  See \code{dwishartc}.
}
\seealso{
  \url{www.bayesian-inference.com/softwaredoc/dist.Wishart.Cholesky}
}
\keyword{Distribution}
