\name{interval}
\alias{interval}
\title{Constrain to Interval}
\description{
  This is a C++ version of the \code{interval} function in the
  \pkg{LaplacesDemon} package.
}
\usage{
  interval(x, a=-Inf, b=Inf, reflect=TRUE)
}
\arguments{
  \item{x}{See \code{interval}.}
  \item{a}{See \code{interval}.}
  \item{b}{See \code{interval}.}
  \item{reflect}{See \code{interval}.}
}
\details{
 For more information, see the \code{interval} function in the
 \pkg{LaplacesDemon} package.
}
\value{
  See \code{interval}.
}
\author{Statisticat, LLC. \email{software@bayesian-inference.com}}
\seealso{
  \url{www.bayesian-inference.com/softwaredoc/interval}
}
\keyword{Utility}
 
