\name{LaplacesDemonCpp-package}
\alias{LaplacesDemonCpp-package}
\alias{LaplacesDemonCpp}
\alias{.colVars}
\alias{.iqagh}
\alias{.iqaghsg}
\alias{.iqcagh}
\alias{.laaga}
\alias{.labfgs}
\alias{.labhhh}
\alias{.lacg}
\alias{.ladfp}
\alias{.lahar}
\alias{.lahj}
\alias{.lalbfgs}
\alias{.lalm}
\alias{.lanm}
\alias{.lanr}
\alias{.lapso}
\alias{.larprop}
\alias{.lasgd}
\alias{.lasoma}
\alias{.laspg}
\alias{.lasr1}
\alias{.latr}
\alias{.mcmcadmg}
\alias{.mcmcagg}
\alias{.mcmcahmc}
\alias{.mcmcaies}
\alias{.mcmcam}
\alias{.mcmcamm}
\alias{.mcmcamm.b}
\alias{.mcmcamwg}
\alias{.mcmccharm}
\alias{.mcmcdemc}
\alias{.mcmcdram}
\alias{.mcmcdrm}
\alias{.mcmcess}
\alias{.mcmcgibbs}
\alias{.mcmcgg}
\alias{.mcmcggcp}
\alias{.mcmcggcpp}
\alias{.mcmcggdp}
\alias{.mcmcggdpp}
\alias{.mcmcharm}
\alias{.mcmchmc}
\alias{.mcmchmcda}
\alias{.mcmcim}
\alias{.mcmcinca}
\alias{.mcmcmala}
\alias{.mcmcmcmcmc}
\alias{.mcmcmtm}
\alias{.mcmcmwg}
\alias{.mcmcnuts}
\alias{.mcmcohss}
\alias{.mcmcram}
\alias{.mcmcrdmh}
\alias{.mcmcrefractive}
\alias{.mcmcrj}
\alias{.mcmcrss}
\alias{.mcmcrwm}
\alias{.mcmcsamwg}
\alias{.mcmcsgld}
\alias{.mcmcslice}
\alias{.mcmcsmwg}
\alias{.mcmcthmc}
\alias{.mcmctwalk}
\alias{.mcmcuess}
\alias{.mcmcusamwg}
\alias{.mcmcusmwg}
\alias{.rowVars}
\alias{.vbsalimans2}
\docType{package}
\title{
  LaplacesDemonCpp is an R package that includes C++ functions as an
  extension to the LaplacesDemon package.
}
\description{
  The LaplacesDemon package is a complete environment for Bayesian
  inference in R. LaplacesDemon is written entirely in R. C++ is
  valuable for speed. This package consists of C++ functions to be used
  with the LaplacesDemon package.
}
\details{
  \tabular{ll}{
    Package: \tab LaplacesDemonCpp\cr
    Type: \tab Package\cr
    Version: \tab 2014-05-12\cr
    Date: \tab 2014-05-12\cr
    License: \tab MIT\cr
  }
  The LaplacesDemonCpp package consists of C++ functions that replace
  functions written in R in LaplacesDemon.

  Documentation exists, but is minimal, because these functions are
  fully documented in the LaplacesDemon package.

  To use LaplacesDemonCpp, simply activate the LaplacesDemonCpp library,
  and any C++ functions here will replace the R functions in
  LaplacesDemon.
}
\author{Statisticat, LLC. \email{software@bayesian-inference.com}}
\keyword{package}
