\name{Math}
\alias{partial}
\title{Math Utility Functions}
\description{
  These are C++ versions of utility functions for math.
}
\usage{
partial(Model, parm, Data, Interval=1e-6, Method="simple")
}
\arguments{
  \item{Model}{See \code{partial}.}
  \item{parm}{See \code{partial}.}
  \item{Data}{See \code{partial}.}
  \item{Interval}{See \code{partial}.}
  \item{Method}{See \code{partial}.}
}
\details{
  For more information, see the \code{partial} function in the
 \pkg{LaplacesDemon} package.
}
\value{
  See \code{partial}.
}
\author{Statisticat, LLC. \email{software@bayesian-inference.com}}
\seealso{
  \url{www.bayesian-inference.com/softwaredoc/Math}
}
\keyword{Math}