\name{Matrices}
\alias{is.symmetric.matrix}
\alias{tr}
\title{Matrix Utility Functions}
\description{
  These are C++ versions of utility functions for working with matrices.
}
\usage{
is.symmetric.matrix(x)
tr(x)
}
\arguments{
  \item{x}{See \code{is.symmetric.matrix}.}
}
\details{
  See \code{is.symmetric.matrix}.
}
\author{Statisticat, LLC. \email{software@bayesian-inference.com}}
\seealso{
  \url{www.bayesian-inference.com/softwaredoc/Matrices}
}
\keyword{Array}
\keyword{High Performance Computing}
\keyword{Math}
\keyword{Matrix}
\keyword{Utility}