\name{dist.HalfCauchy}
\alias{dhalfcauchy}
\alias{rhalfcauchy}
\title{Half-Cauchy Distribution}
\description{
  These are the C++ versions of the \code{dhalfcauchy} and
  \code{rhalfcauchy} functions in the \pkg{LaplacesDemon} package.
}
\usage{
dhalfcauchy(x, scale=25, log=FALSE)
rhalfcauchy(n, scale=25)
}
\arguments{
  \item{x}{See \code{dhalfcauchy}.}
  \item{n}{See \code{rhalfcauchy}.}
  \item{scale}{See \code{dhalfcauchy}.}
  \item{log}{See \code{dhalfcauchy}.}
}
\details{
  See \code{dhalfcauchy}.
}
\value{
  See \code{dhalfcauchy}.
}
\seealso{
  \url{www.bayesian-inference.com/softwaredoc/dist.HalfCauchy}
}
\keyword{Distribution}
