\name{LaplacesDemonCpp-package}
\alias{LaplacesDemonCpp-package}
\alias{LaplacesDemonCpp}
\docType{package}
\title{
  LaplacesDemonCpp is an R package that includes C++ functions as an
  extension to the LaplacesDemon package.
}
\description{
  The LaplacesDemon package is a complete environment for Bayesian
  inference in R. LaplacesDemon is written entirely in R. C++ is
  valuable for speed. This package consists of C++ functions to be used
  with the LaplacesDemon package.
}
\details{
  \tabular{ll}{
    Package: \tab LaplacesDemonCpp\cr
    Type: \tab Package\cr
    Version: \tab 2014-05-12\cr
    Date: \tab 2014-05-12\cr
    License: \tab MIT\cr
  }
  The LaplacesDemonCpp package consists of C++ functions that replace
  functions written in R in LaplacesDemon.

  Documentation exists, but is minimal, because these functions are
  fully documented in the LaplacesDemon package.

  To use LaplacesDemonCpp, simply activate the LaplacesDemonCpp library,
  and any C++ functions here will replace the R functions in
  LaplacesDemon.
}
\author{Statisticat, LLC. \email{software@bayesian-inference.com}}
\keyword{package}
